Fixed issue #370

+++ src/seq/fasta_util.cpp
@@ -105,7 +105,7 @@ void generateDbBlock(const std::string &db_file_name, int block_size,
   // get total_size;
   FastaReader size_reader(db_file_name);
   FastaSeqPtr seq_info = size_reader.getNextSeq();
-  long total_seq_size = 0;
+  double total_seq_size = 0;
   int seq_num = 0;
   while (seq_info != nullptr) {
     std::string seq = seq_info->getRawSeq();
@@ -121,6 +121,7 @@ void generateDbBlock(const std::string &db_file_name, int block_size,
     seq_info = size_reader.getNextSeq();
   }
   size_reader.close();
+  //LOG_ERROR("total seq size " << total_seq_size); 
   // adjust block size if the database is too small and the sequence number 
   // is large enough > 50 * min_block_num
   int block_num = static_cast<int>(std::ceil(total_seq_size/block_size));
